requirements.txt
asgiref==3.4.1
Django==3.1.14
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-multiselectfield==0.1.12
django-taggit==2.0.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
mysqlclient==2.1.0
Pillow==8.3.2
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
